third_party/blink/web_tests/platform/generic/fast/dom/Window/alert-undefined-expected.txt

ALERT: undefined
ALERT: null
CONFIRM: 
CONFIRM: null
PROMPT: , default text: 
PROMPT: null, default text: 
ALERT: 
CONFIRM: 
PROMPT: , default text: 
